requirements.txt
streamlit
python-telegram-bot==21.0.1
PyMuPDF
Pillow
python-dotenv
